try and reenable qsort for 386 (again)

+++ func_test.go
@@ -74,10 +74,10 @@ func Test_qsort(t *testing.T) {
 
 	data := []int{88, 56, 100, 2, 25}
 	sorted := []int{2, 25, 56, 88, 100}
-	compare := func(a, b *int) int32 {
-		return int32(*a - *b)
+	compare := func(a, b *int) uintptr {
+		return uintptr(*a - *b)
 	}
-	var qsort func(data []int, nitms uintptr, size uintptr, compar func(a, b *int) int32)
+	var qsort func(data []int, nitms uintptr, size uintptr, compar func(a, b *int) uintptr)
 	purego.RegisterLibFunc(&qsort, libc, "qsort")
 	qsort(data, uintptr(len(data)), unsafe.Sizeof(int(0)), compare)
 	for i := range data {
